fix(xtask): discover repository root at runtime
